* Move PinDialog that checks pin upon startup from Homepage to Assets page.
Fix: Correctly handle simultaneous account and network change:
Before: PinDialog was shown to early while the network was still loading. This lead to it being dissmissed after the network loaded.

Solution: Before displaying the dialog, check if Assets is the currentRoute such that the Dialog is shown after the network was loaded.