flutter format . --line-length 120
implementing comments in conversation of #202,
no more child widget updating parent.
correct dynamic string insertion with i18n library

cleanup

bugfix: when switching to accountSelector and choosing an account, the old passwordDialog with the former accountname was shown.
Solution was to pop the context and then to go to accountSelection.
But then when switching back and forth accountSelection and Index/Assets, the parent widget (assets) wouldn't rebuild for a while and therefore not show any passwordDialog --> unacceptable.
So pass a callback function to the child widget to setState in the parent widget s.th. rebuilding the parent widget is guaranteed no matter what.
Please feedback for best practice, we need to discuss this!

Refactoring, now there is a passwordInputDialog to enter pin or cancel AND a passwordInputSwitchDialog to enter password or switch users.

removed the functionality of resetting all accounts.
Now instead we can switch accounts if password forgotten and only delete an account if the pin is known.
The cancel button became optional, so everytime a passwordInputDialog is opened, the OnCancel has to be defined.

added todo

remove comments

fixing the removeAllAccounts function. Needs to be async, else, not all accounts are deleted

refactoring passwordInputDialog to only show Reset Button, if onReset defined. Thx clang!

And from now on: flutter format on save is on

refactoring removingAllAccounts and cleanUp

removing unnecessary prints

removing unnecessary prints

fixing lint issues

fixed cancel issue, now nagivation occurs as wanted

enabling resetting all accounts, when user doesn't know pin
instead of removing app and reinstalling

in accountManagePage, after delete option, possiblity to reset all accounts and go back to main page

added reset to passwordInputDialog

fix error, that after account deletion a check in a non existing list is made.

removing all accounts if pin forgotten
for now when managing accounts and removing an account, maybe this should also be an option?